# back-end/openapi.yaml
openapi: 3.0.3
info:
  title: Accounting API
  version: 0.2.0
  description: |
    English: Phase 1 auth endpoints with JWT access/refresh tokens.
    فارسی: فاز ۱؛ اندپوینت‌های احراز هویت با توکن دسترسی و رفرش.
  x-i18n:
    description:
      en: Phase 1 auth endpoints with JWT access/refresh tokens.
      fa: فاز ۱؛ اندپوینت‌های احراز هویت با توکن دسترسی و رفرش.
servers:
  - url: http://localhost:4100/api
paths:
...
  /v1/journals:
    get:
      summary: List journals
      x-i18n:
        summary:
          en: List journals
          fa: فهرست سندها
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of journals
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Journal'
                  message:
                    type: string
    post:
      summary: Create journal draft
      x-i18n:
        summary:
          en: Create journal draft
          fa: ایجاد پیش‌نویس سند
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fiscal_year_id:
                  type: string
                date:
                  type: string
                  format: date
                ref_no:
                  type: string
                  nullable: true
                description:
                  type: string
                  nullable: true
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/JournalItemInput'
              required: [fiscal_year_id, date, items]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum: [draft, posted]
                  message:
                    type: string
  /v1/journals/{id}:
    get:
      summary: Get journal
      x-i18n:
        summary:
          en: Get journal
          fa: دریافت سند
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Journal with items
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/Journal'
                  message:
                    type: string
        '404':
          description: Not found
    patch:
      summary: Update journal (draft only)
      x-i18n:
        summary:
          en: Update journal (draft only)
          fa: به‌روزرسانی سند (فقط پیش‌نویس)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                ref_no:
                  type: string
                  nullable: true
                description:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
    delete:
      summary: Delete journal (draft only)
      x-i18n:
        summary:
          en: Delete journal (draft only)
          fa: حذف سند (فقط پیش‌نویس)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
        '404':
          description: Not found
  /v1/journals/{id}/post:
    post:
      summary: Post journal
      x-i18n:
        summary:
          en: Post journal
          fa: پست کردن سند
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Posted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum: [draft, posted]
                  message:
                    type: string
        '404':
          description: Not found
...
  /v1/journals/{id}/reverse:
    post:
      summary: Reverse posted journal
      x-i18n:
        summary:
          en: Reverse posted journal
          fa: برگشت سند پست‌شده
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reversed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum: [posted]
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Journal:
      type: object
      properties:
        id:
          type: string
        fiscal_year_id:
          type: string
        ref_no:
          type: string
          nullable: true
        date:
          type: string
          format: date
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, posted]
        items:
          type: array
          items:
            $ref: '#/components/schemas/JournalItem'
      required: [id, fiscal_year_id, date, status]
    JournalItem:
      type: object
      properties:
        id:
          type: string
        journal_id:
          type: string
        account_id:
          type: string
        party_id:
          type: string
          nullable: true
        debit:
          type: number
        credit:
          type: number
        description:
          type: string
          nullable: true
      required: [id, journal_id, account_id, debit, credit]
    JournalItemInput:
      type: object
      properties:
        account_id:
          type: string
        party_id:
          type: string
          nullable: true
        debit:
          type: number
        credit:
          type: number
        description:
          type: string
          nullable: true
      required: [account_id, debit, credit]
...